% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmle_prev.R
\name{atmle}
\alias{atmle}
\title{Adaptive-TMLE for RCT-ATE}
\usage{
atmle(
  data,
  S,
  W,
  A,
  Y,
  controls_only,
  family,
  theta_method = "glmnet",
  Pi_method = "glmnet",
  g_method = "glmnet",
  g_delta_method = "glmnet",
  theta_tilde_method = "glmnet",
  bias_working_model = "HAL",
  bias_working_model_formula = NULL,
  pooled_working_model = "HAL",
  pooled_working_model_formula = NULL,
  cross_fit_nuisance = TRUE,
  min_working_model = FALSE,
  max_degree = 1,
  v_folds = NULL,
  stratify = TRUE,
  g_bounds = NULL,
  Pi_bounds = NULL,
  theta_bounds = NULL,
  target_gwt = TRUE,
  verbose = TRUE,
  max_iter = 50,
  target_method = "relaxed",
  eic_method = "svd_pseudo_inv",
  alpha = 0.05,
  enumerate_basis_args = list(max_degree = 2, smoothness_orders = 1, num_knots = c(20,
    5)),
  browse = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing baseline covariates \eqn{W},
binary treatment indicator \eqn{A} (\eqn{A=1} for active treatment),
outcome \eqn{Y}, and binary study indicator of whether the observation is
from the randomized controlled trial \eqn{S=1} or from the external data
\eqn{S=0}. If both studies are observational, then \eqn{S=1} should be the
reference study.}

\item{controls_only}{A logical indicating whether the external data has
only control-arm or both control-arm and treatment-arm.}

\item{family}{A character string specifying the family of the outcome
\eqn{Y}. Currently only \code{"gaussian"} and \code{"binomial"} are
supported.}

\item{theta_method}{The method to estimate the nuisance function
\eqn{\theta(W,A)=\mathbb{E}(Y\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}.}

\item{Pi_method}{The method to estimate the nuisance function
\eqn{\Pi(S\mid W,A)=\mathbb{P}(S=1\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}.}

\item{g_method}{The method to estimate the nuisance function
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}.}

\item{g_delta_method}{The method to estimate the nuisance function
\eqn{\delta(W,A)=\mathbb{P}(\Delta=1\mid W,A)}. \code{"glm"} for main-term
linear model, \code{"glmnet"} for lasso, \code{"sl3"} for default super
learner, or a \code{list} of \code{sl3} learners. Default is \code{"glmnet"}.
\eqn{\Delta} is the indicator of missing outcome. \code{1} - observed,
\code{0} - missing. Only applicable when there are missing outcomes.}

\item{theta_tilde_method}{The method to estimate the nuisance function
\eqn{\tilde{\theta}(W,A)=\mathbb{E}(Y\mid W,A,S=1)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}.}

\item{bias_working_model}{The working model for the bias estimand.
Either \code{"glmnet"} for lasso-based working model or \code{"HAL"} for
highly adaptive lasso-based working model}

\item{pooled_working_model}{The working model for the pooled-ATE estimand.
Either \code{"glmnet"} for lasso-based working model or \code{"HAL"} for
highly adaptive lasso-based working model.}

\item{v_folds}{The number of folds for cross-validation (whenever necessary).
Default is \code{5}.}

\item{g_bounds}{A numeric vector of lower and upper bounds for the
treatment mechanism. The first element is the lower bound, and the second
element is the upper bound. Default is \code{c(0.01, 0.99)}.}

\item{Pi_bounds}{A numeric vector of lower and upper bounds for the
trial enrollment probabilities. The first element is the lower bound,
and the second element is the upper bound. Default is \code{c(0.01, 0.99)}.}

\item{theta_bounds}{A numeric vector of lower and upper bounds for the
conditional mean of outcome given baseline covariates and treatment.
The first element is the lower bound, and the second element is the upper
bound. Default is \code{c(-Inf, Inf)}.}

\item{target_gwt}{If \code{TRUE}, the treatment mechanism is moved from the
denominator of the clever covariate to the weight when fitting the TMLE
submodel.}

\item{verbose}{A logical indicating whether to print out the progress.
Default is \code{TRUE}.}

\item{S_node}{The column index of the \eqn{S} node in \code{data}.}

\item{W_node}{The column indices of the \eqn{W} node in \code{data}.}

\item{A_node}{The column index of the \eqn{A} node in \code{data}.}

\item{Y_node}{The column index of the \eqn{Y} node in \code{data}.}

\item{g_rct}{The probability of receiving the active treatment in the
randomized controlled trial.}

\item{atmle_pooled}{A logical indicating whether to also use A-TMLE for the
pooled-ATE estimand If set to \code{FALSE}, use a regular TMLE for the
pooled-ATE, but A-TMLE for the bias-estimand.
Default is \code{TRUE}.}

\item{Q_method}{The method to estimate the nuisance function
\eqn{Q(A,W)=\mathbb{E}(Y\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
\code{"sl3"} for default super learner, or a \code{list} of \code{sl3}
learners. Default is \code{"glmnet"}. Only applicable when \code{atmle_pooled}
is set to \code{FALSE}.}
}
\value{
A \code{list} containing the following elements:
\item{ate}{The estimated average treatment effect;}
\item{lower}{The lower bound of the \eqn{95\%} confidence interval for the
average treatment effect;}
\item{upper}{The upper bound of the \eqn{95\%} confidence interval for the
average treatment effect;}
}
\description{
Adaptive-TMLE (A-TMLE) for estimating the average treatment
effect based on combined randomized controlled trial data and real-world
data.
}
\examples{
set.seed(123)

n <- 2000
S <- rbinom(n, 1, 0.2)
W1 <- rnorm(n)
W2 <- rnorm(n)
W3 <- rnorm(n)
A <- numeric(n)
g_rct <- 0.67
A[S == 1] <- rbinom(sum(S), 1, g_rct)
A[S == 0] <- rbinom(n - sum(S), 1, plogis(0.5 * W1[S == 0]))
UY <- rnorm(n, 0, 1)
Y <- 2.5 + 0.9 * W1 + 1.1 * W2 + 2.7 * W3 + 1.5 * A + UY + (1 - S) * (0.2 + 0.1 * W1 * (1 - A))
data <- data.frame(S, W1, W2, A, Y)
true_ate <- 1.5

res <- atmle(data,
  S_node = c(1),
  W_node = c(2, 3),
  A_node = 4,
  Y_node = 5,
  controls_only = FALSE,
  family = "gaussian",
  atmle_pooled = TRUE,
  g_rct = g_rct,
  verbose = FALSE
)
}
