% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmle_ate.R
\name{atmle_ate}
\alias{atmle_ate}
\title{Adaptive-TMLE for ATE}
\description{
Adaptive-TMLE for ATE

Adaptive-TMLE for ATE
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-A-TMLE ATE-new}{\code{atmle_ate$new()}}
\item \href{#method-A-TMLE ATE-create_folds}{\code{atmle_ate$create_folds()}}
\item \href{#method-A-TMLE ATE-fit_regression}{\code{atmle_ate$fit_regression()}}
\item \href{#method-A-TMLE ATE-fit_cate}{\code{atmle_ate$fit_cate()}}
\item \href{#method-A-TMLE ATE-augment_wm}{\code{atmle_ate$augment_wm()}}
\item \href{#method-A-TMLE ATE-target}{\code{atmle_ate$target()}}
\item \href{#method-A-TMLE ATE-target_tmle}{\code{atmle_ate$target_tmle()}}
\item \href{#method-A-TMLE ATE-target_relaxed}{\code{atmle_ate$target_relaxed()}}
\item \href{#method-A-TMLE ATE-inference}{\code{atmle_ate$inference()}}
\item \href{#method-A-TMLE ATE-run_init_est}{\code{atmle_ate$run_init_est()}}
\item \href{#method-A-TMLE ATE-run}{\code{atmle_ate$run()}}
\item \href{#method-A-TMLE ATE-clone}{\code{atmle_ate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-new"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$new(
  data,
  W_nodes,
  A_node,
  Y_node,
  family,
  n_folds = NULL,
  seed = 123
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-create_folds"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-create_folds}{}}}
\subsection{Method \code{create_folds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$create_folds(n_eff, n_folds, strata_ids, seed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-fit_regression"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-fit_regression}{}}}
\subsection{Method \code{fit_regression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$fit_regression(
  method,
  folds,
  covariate_nodes,
  outcome_node,
  weight_node = NULL,
  subset = seq(nrow(self$data)),
  bound = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-fit_cate"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-fit_cate}{}}}
\subsection{Method \code{fit_cate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$fit_cate(hal_args, parallel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-augment_wm"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-augment_wm}{}}}
\subsection{Method \code{augment_wm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$augment_wm(target_method, parallel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-target"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-target}{}}}
\subsection{Method \code{target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$target(n_lambda, method)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-target_tmle"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-target_tmle}{}}}
\subsection{Method \code{target_tmle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$target_tmle(beta, phi_W)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-target_relaxed"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-target_relaxed}{}}}
\subsection{Method \code{target_relaxed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$target_relaxed(pseudo_outcome, pseudo_weights, phi_W)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-inference"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-inference}{}}}
\subsection{Method \code{inference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$inference(
  alpha = 0.05,
  small_diag = 0.001,
  fall_back_method = "svd_pseudo_inv"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-run_init_est"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-run_init_est}{}}}
\subsection{Method \code{run_init_est()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$run_init_est(
  theta_method,
  g_method,
  learn_theta_via_Q = TRUE,
  g_bound = NULL,
  verbose = FALSE,
  browse = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-run"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$run(
  theta_method,
  g_method,
  family,
  learn_theta_via_Q = TRUE,
  n_lambda = 1,
  cate_hal_args = list(max_degree = 3L, smoothness_orders = 1L, num_knots = 20L),
  target_method = "tmle",
  parallel = FALSE,
  browse = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE ATE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE ATE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
