% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ci.R
\name{bootstrap_psi_pound}
\alias{bootstrap_psi_pound}
\title{Function to compute bootstrap confidence interval for the bias
estimate.}
\usage{
bootstrap_psi_pound(tau, W, Pi, B = 1000)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{Pi}{A \code{list} containing the following elements:
\item{pred}{The estimated trial enrollment probabilities;}
\item{A1}{The estimated trial enrollment probabilities under treatment;}
\item{A0}{The estimated trial enrollment probabilities under control.}}

\item{B}{The number of bootstrap samples.}

\item{A}{\code{list} containing the following elements:
\item{A1}{A numeric vector of the estimated counterfactual conditional
effects under treatment;}
\item{A0}{A numeric vector of the estimated counterfactual conditional
effects under control;}
\item{x_basis}{A numeric matrix of the working model bases;}
\item{x_basis_A1}{A numeric matrix of the counterfactual working model bases
under treatment;}
\item{x_basis_A0}{A numeric matrix of the counterfactual working model bases
under control;}
\item{pred}{A numeric vector of estimated conditional effects;}
\item{coefs}{A numeric vector of the working model coefficients.}}
}
\description{
Function to compute bootstrap confidence interval for the bias
estimate.
}
\keyword{internal}
