% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ci.R
\name{bootstrap_psi_tilde}
\alias{bootstrap_psi_tilde}
\title{Function to compute bootstrap confidence interval for
the pooled-ATE estimate.}
\usage{
bootstrap_psi_tilde(W, psi_tilde, B = 1000)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{psi_tilde}{A \code{list} containing the following elements:
\item{pred}{A numeric vector of the estimated conditional effects;}
\item{x_basis}{A numeric matrix of the working model bases;}
\item{coefs}{A numeric vector of the working model coefficients.}}

\item{B}{The number of bootstrap samples.}
}
\description{
Function to compute bootstrap confidence interval for
the pooled-ATE estimate.
}
