% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_args}
\alias{check_args}
\title{Sanity checks for other arguments}
\usage{
check_args(controls_only, atmle_pooled, var_method, g_rct, verbose)
}
\description{
Sanity checks for other arguments
}
