% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_data_and_nodes}
\alias{check_data_and_nodes}
\title{Sanity checks for input data and nodes}
\usage{
check_data_and_nodes(data, S_node, W_node, A_node, Y_node)
}
\arguments{
\item{data}{A \code{data.frame} containing baseline covariates \eqn{W}, binary
treatment indicator \eqn{A} (\eqn{A=1} for active treatment),
outcome \eqn{Y}, and binary indicator of whether the observation is from the
randomized controlled trial \eqn{S=1} or from the external data \eqn{S=0}.}

\item{S_node}{The column indices of the \eqn{S} node in \code{data}.}

\item{W_node}{The column indices of the \eqn{W} node in \code{data}.}

\item{A_node}{The column indices of the \eqn{A} node in \code{data}.}

\item{Y_node}{The column indices of the \eqn{Y} node in \code{data}.}
}
\description{
Sanity checks for input data and nodes
}
\keyword{internal}
