% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_learners}
\alias{check_learners}
\title{Sanity checks for learners}
\usage{
check_learners(
  theta_method,
  Pi_method,
  g_method,
  theta_tilde_method,
  Q_method,
  bias_working_model,
  pooled_working_model
)
}
\description{
Sanity checks for learners
}
