% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{eic_ate_atmle}
\alias{eic_ate_atmle}
\title{Efficient influence function for the bias projection parameter}
\usage{
eic_ate_atmle(
  Y,
  A,
  g1W,
  theta,
  phi_W,
  cate_pred,
  small_diag = 0.001,
  fall_back_method = "svd_pseudo_inv"
)
}
\arguments{
\item{Y}{A vector of outcomes.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{theta}{Results from \code{\link{learn_theta}}.}

\item{Pi}{Results from \code{\link{learn_Pi}}.}

\item{tau}{Results from \code{\link{learn_tau}}.}

\item{g}{Results from \code{\link{learn_g}}.}

\item{psi_pound_est}{Estimate for the bias estimand.}

\item{S}{A vector of study indicators, \eqn{S=1} for RCT, \eqn{S=0} for RWD.}

\item{n}{The number of observations.}

\item{controls_only}{A logical indicating whether external data has only
control-arm observations.}

\item{weights}{A vector of (e.g. inverse-censoring) weights.}
}
\value{
A vector of efficient influence function values.
}
\description{
Computes the efficient influence function for the bias
projection parameter.
}
\keyword{internal}
