% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{get_eic_psi_tilde}
\alias{get_eic_psi_tilde}
\title{Efficient influence function for the pooled-ATE parameter under the
working model.}
\usage{
get_eic_psi_tilde(psi_tilde, g_pred, theta, Y, A, n)
}
\arguments{
\item{psi_tilde}{Returns from \code{\link{learn_psi_tilde}}.}

\item{g_pred}{Returns from \code{\link{learn_g}}.}

\item{theta}{Returns from \code{\link{learn_theta}}.}

\item{Y}{A vector of outcomes.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{n}{The number of observations.}
}
\value{
A vector of efficient influence function values.
}
\description{
Computes the efficient influence function for the pooled-ATE
parameter under the working model.
}
\keyword{internal}
