% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_g.R
\name{learn_g}
\alias{learn_g}
\title{Learn nuisance function: treatment mechanism}
\usage{
learn_g(S, W, A, g_rct, controls_only, method, v_folds)
}
\arguments{
\item{S}{A vector of study indicators, \eqn{S=1} for RCT, \eqn{S=0} for RWD.}

\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{g_rct}{A numeric of treatment probability in RCT.}

\item{controls_only}{A logical indicating whether to learn only among
controls. This applies when the external data only has control-arm
observations.}

\item{method}{Learning method. \code{"glm"} for main-term linear model,
\code{"glmnet"} for lasso, or a \code{list} of \code{sl3} learners for
super learner-based estimation.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}
}
\value{
A numeric vector of estimated treatment probabilities.
}
\description{
Function to learn the treatment mechanism
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.
}
