% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_g_delta_tilde.R
\name{learn_g_delta_tilde}
\alias{learn_g_delta_tilde}
\title{Learn nuisance function: conditional probability of observing the
outcome given baseline covariates and treatment}
\usage{
learn_g_delta_tilde(
  W,
  A,
  delta,
  method,
  folds,
  g_bounds,
  cross_fit_nuisance = TRUE
)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators.}

\item{delta}{A vector of missingness indicators.}

\item{method}{Learning method. \code{"glm"} for main-term linear model,
\code{"glmnet"} for lasso, or a \code{list} of \code{sl3} learners for
super learner-based estimation.}

\item{g_bounds}{A numeric vector of lower and upper bounds for the
missingness mechanism. The first element is the lower bound, and the second
element is the upper bound.}
}
\value{
A numeric vector of the estimated values.
}
\description{
Function to learn the conditional probability of observing the
outcome given baseline covariates and treatment,
\eqn{\tilde{g}_\Delta(1\mid W,A)=\mathbb{P}(\Delta=1\mid W,A)}. Only applicable when
the outcome is subject to missingness.
}
\keyword{nuisance}
