% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_Pi.R
\name{learn_Pi}
\alias{learn_Pi}
\title{Learn nuisance function: trial enrollment mechanism}
\usage{
learn_Pi(S, W, A, controls_only, method, v_folds)
}
\arguments{
\item{S}{A vector of study indicators, \eqn{S=1} for RCT, \eqn{S=0} for RWD.}

\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{controls_only}{A logical indicating whether to learn only among
controls. This applies when the external data only has control-arm
observations.}

\item{method}{Learning method. \code{"glm"} for main-term linear model,
\code{"glmnet"} for lasso, or a \code{list} of \code{sl3} learners for
super learner-based estimation.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}
}
\value{
A \code{list} containing the following elements:
\item{pred}{The estimated trial enrollment probabilities;}
\item{A1}{The estimated trial enrollment probabilities under treatment;}
\item{A0}{The estimated trial enrollment probabilities under control.}
}
\description{
Function to learn the trial enrollment mechanism
\eqn{\Pi(S\mid W,A)=\mathbb{P}(S=1\mid W,A)}.
}
