% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_Q.R
\name{learn_Q}
\alias{learn_Q}
\title{Learn outcome regression}
\usage{
learn_Q(W, A, Y, method, v_folds, family, theta_bounds)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{Y}{A vector of outcomes.}

\item{method}{Learning method. \code{"glm"} for main-term linear model,
\code{"glmnet"} for lasso, or a \code{list} of \code{sl3} learners for
super learner-based estimation.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}

\item{family}{A character string specifying the family of the outcome
\eqn{Y}. Either \code{"gaussian"} or \code{"binomial"}.}

\item{theta_bounds}{A numeric vector of lower and upper bounds for the
conditional mean of outcome given baseline covariates and treatment.
The first element is the lower bound, and the second element is the upper
bound.}
}
\value{
A numeric vector of estimated values.
}
\description{
Function to learn the outcome regression
\eqn{Q(A,W)=\mathbb{E}(Y\mid A,W)}.
}
\keyword{nuisance}
