% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_T.R
\name{learn_T}
\alias{learn_T}
\title{Learn working model for the conditional effect of treatment on the outcome}
\usage{
learn_T(W, A, Y, g, theta_tilde, method, v_folds)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{Y}{A vector of outcomes.}

\item{g}{A vector of estimated treatment probabilities,
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.}

\item{theta_tilde}{A vector of estimated conditional mean of outcome given
baseline covariates, \eqn{\theta(W)=\mathbb{E}(Y\mid W)}.}

\item{method}{Working model type. Either \code{"glmnet"} for lasso-based
working model or \code{"HAL"} for highly adaptive lasso-based working model.
Default is \code{"glmnet"}.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}
}
\value{
A \code{list} containing the following elements:
\item{pred}{A numeric vector of the estimated conditional effects;}
\item{x_basis}{A numeric matrix of the working model bases;}
\item{coefs}{A numeric vector of the working model coefficients.}
}
\description{
Function to learn the conditional effect of treatment on the
outcome, \eqn{T(W)=\mathbb{E}(Y\mid W,A=1)-\mathbb{E}(Y\mid W,A=0)}.
}
