% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_tau.R
\name{learn_tau}
\alias{learn_tau}
\title{Learn working model for the conditional effect of the study indicator \
on the outcome}
\usage{
learn_tau(S, W, A, Y, Pi, theta, controls_only, method, v_folds, max_degree)
}
\arguments{
\item{S}{A vector of study indicators, \eqn{S=1} for RCT, \eqn{S=0} for RWD.}

\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{Y}{A vector of outcomes.}

\item{Pi}{A vector of estimated trial enrollment probabilities,
\eqn{\Pi(W,A)=\mathbb{P}(S=1\mid W,A)}.}

\item{theta}{A vector of estimated conditional mean of outcome given
baseline covariates and treatment, \eqn{\theta(W,A)=\mathbb{E}(Y\mid W,A)}.}

\item{controls_only}{A logical indicating whether the external data has only
control-arm observations.}

\item{method}{Working model type. Either \code{"glmnet"} for lasso-based
working model or \code{"HAL"} for highly adaptive lasso-based working model.
Default is \code{"glmnet"}.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}
}
\value{
A \code{list} containing the following elements:
\item{A1}{A numeric vector of the estimated counterfactual conditional
effects under treatment;}
\item{A0}{A numeric vector of the estimated counterfactual conditional
effects under control;}
\item{x_basis}{A numeric matrix of the working model bases;}
\item{x_basis_A1}{A numeric matrix of the counterfactual working model bases
under treatment;}
\item{x_basis_A0}{A numeric matrix of the counterfactual working model bases
under control;}
\item{pred}{A numeric vector of estimated conditional effects;}
\item{coefs}{A numeric vector of the working model coefficients.}
}
\description{
Function to learn the conditional effect of the study indicator
on the outcome,
\eqn{\tau(W,A)=\mathbb{E}(Y\mid S=1,W,A)-\mathbb{E}(Y\mid S=0,W,A)}.
}
\keyword{model}
\keyword{working}
