% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_tau_A.R
\name{learn_tau_A}
\alias{learn_tau_A}
\title{Learn working model for the conditional effect of treatment on the
outcome}
\usage{
learn_tau_A(
  W,
  A,
  Y,
  g1W,
  delta,
  theta_W,
  method,
  foldid,
  weights,
  enumerate_basis_args,
  pooled_working_model_formula,
  target_method = "oneshot",
  eic_method = "svd_pseudo_inv",
  verbose = FALSE,
  browse = FALSE
)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{Y}{A vector of outcomes.}

\item{method}{Working model type. Either \code{"glmnet"} for lasso-based
working model or \code{"HAL"} for highly adaptive lasso-based working model.
Default is \code{"glmnet"}.}

\item{g}{A vector of estimated treatment probabilities,
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.}

\item{theta_W_tilde}{A vector of estimated conditional mean of outcome given
baseline covariates, \eqn{\theta_W(W)=\mathbb{E}(Y\mid W)}.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}
}
\value{
A \code{list} containing the following elements:
\item{pred}{A numeric vector of the estimated conditional effects;}
\item{x_basis}{A numeric matrix of the working model bases;}
\item{coefs}{A numeric vector of the working model coefficients.}
}
\description{
Function to learn the conditional effect of treatment on the
outcome, \eqn{T(W)=\mathbb{E}(Y\mid W,A=1)-\mathbb{E}(Y\mid W,A=0)}.
}
\keyword{model}
\keyword{working}
