% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_theta.R
\name{learn_theta}
\alias{learn_theta}
\title{Learn nuisance function: conditional mean of outcome given baseline
covariates and treatment}
\usage{
learn_theta(W, A, Y, controls_only, method, v_folds)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators.}

\item{Y}{A vector of outcomes.}

\item{method}{Learning method. \code{"glm"} for main-term linear model,
\code{"glmnet"} for lasso, or a \code{list} of \code{sl3} learners for
super learner-based estimation.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}
}
\value{
A numeric vector of the estimated values.
}
\description{
Function to learn the conditional mean of outcome given
baseline covariates and treatment,
\eqn{\theta(W,A)=\mathbb{E}(Y\mid W,A)}.
}
