% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_theta_tilde.R
\name{learn_theta_tilde}
\alias{learn_theta_tilde}
\title{Learn nuisance function: conditional mean of outcome given baseline
covariates}
\usage{
learn_theta_tilde(W, Y, method, v_folds, family)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{Y}{A vector of outcomes.}

\item{method}{Learning method. \code{"glm"} for main-term linear model,
\code{"glmnet"} for lasso, or a \code{list} of \code{sl3} learners for
super learner-based estimation.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}

\item{family}{A character string specifying the family of the outcome
\eqn{Y}.}
}
\value{
A numeric vector of the estimated values.
}
\description{
Function to learn the conditional mean of outcome given
baseline covariates, \eqn{\tilde{\theta}(W)=\mathbb{E}(Y\mid W)}.
}
