% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_theta_W.R
\name{learn_theta_W}
\alias{learn_theta_W}
\title{Learn nuisance function: conditional mean of outcome given baseline
covariates}
\usage{
learn_theta_W(
  W,
  Y,
  delta,
  weights,
  method,
  folds,
  family,
  theta_bounds,
  cross_fit_nuisance
)
}
\arguments{
\item{W}{A matrix of baseline covariates.}

\item{Y}{A vector of outcomes.}

\item{method}{Learning method. \code{"glm"} for main-term linear model,
\code{"glmnet"} for lasso, or a \code{list} of \code{sl3} learners for
super learner-based estimation.}

\item{family}{A character string specifying the family of the outcome
\eqn{Y}. Either \code{"gaussian"} or \code{"binomial"}.}

\item{theta_bounds}{A numeric vector of lower and upper bounds for the
conditional mean of outcome given baseline covariates.
The first element is the lower bound, and the second element is the upper
bound.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}
}
\value{
A numeric vector of the estimated values.
}
\description{
Function to learn the conditional mean of outcome given
baseline covariates, \eqn{\tilde{\theta}(W)=\mathbb{E}(Y\mid W)}.
}
\keyword{nuisance}
