% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_counter_design_matrix.R
\name{make_counter_design_matrix}
\alias{make_counter_design_matrix}
\title{Make counterfactual design matrix}
\usage{
make_counter_design_matrix(basis_list, X_counterfactual, X_unpenalized = NULL)
}
\arguments{
\item{basis_list}{A \code{list} of HAL bases.}

\item{X_counterfactual}{A \code{matrix} containing baseline covariates
evaluated at the desired counterfactual.}

\item{X_unpenalized}{A \code{matrix} containing baseline covariates or
bases that are not penalized by HAL. Default is \code{NULL}.}
}
\description{
Make counterfactual design matrix
}
\keyword{internal}
