% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{nonparametric}
\alias{nonparametric}
\title{Nonparametric TMLE}
\usage{
nonparametric(
  data,
  S,
  W,
  A,
  Y,
  controls_only,
  family = "gaussian",
  atmle_pooled = TRUE,
  theta_method = "glm",
  Pi_method = "glm",
  g_method = "glm",
  theta_tilde_method = "glm",
  Q_method = "glm",
  bias_working_model = "glmnet",
  pooled_working_model = "glmnet",
  var_method = "ic",
  max_iter = 1,
  v_folds = 5,
  verbose = TRUE
)
}
\description{
Nonparametric TMLE
}
