% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procova.R
\name{procova}
\alias{procova}
\title{PROCOVA}
\usage{
procova(
  data,
  S_node,
  W_node,
  A_node,
  Y_node,
  controls_only,
  family,
  g_rct,
  Q_method = "glmnet",
  g_method = "glmnet",
  g_delta_method = "glmnet",
  v_folds = 5,
  g_bounds = c(0.01, 0.99),
  target_gwt = TRUE,
  verbose = TRUE
)
}
\description{
PROCOVA
}
