% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHAL.R
\name{rHAL}
\alias{rHAL}
\title{R-learner for Quasi-Oracle Estimation of Heterogeneous Treatment
Effects}
\usage{
rHAL(
  W,
  A,
  Y,
  g1W,
  theta,
  foldid,
  weights = NULL,
  enumerate_basis_args = list(max_degree = 2, smoothness_orders = 1, num_knots = c(20,
    5)),
  use_weight = TRUE,
  browse = FALSE
)
}
\description{
Code is adapted from \url{https://github.com/xnie/rlearner}
}
