% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{rlasso}
\alias{rlasso}
\title{R-learner for Quasi-Oracle Estimation of Heterogeneous Treatment
Effects}
\usage{
rlasso(
  W,
  A,
  Y,
  g1W,
  theta,
  foldid,
  weights = NULL,
  max_degree = 1,
  use_weight = TRUE,
  browse = FALSE
)
}
\description{
Code is adapted from \url{https://github.com/xnie/rlearner}
}
