% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_Pi.R
\name{target_Pi}
\alias{target_Pi}
\title{TMLE update of the trial enrollment probabilities}
\usage{
target_Pi(
  S,
  W,
  A,
  Y,
  delta,
  g1W,
  Pi,
  theta_WA,
  tau_S,
  controls_only,
  target_gwt,
  weights,
  Pi_bounds
)
}
\arguments{
\item{S}{A vector of study indicators, \eqn{S=1} for RCT, \eqn{S=0} for RWD.}

\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{Pi}{Returns from the \code{\link{learn_Pi}} function.}

\item{controls_only}{A logical indicating whether the external data has only
control-arm observations.}

\item{target_gwt}{If \code{TRUE}, the treatment mechanism is moved from the
denominator of the clever covariate to the weight when fitting the TMLE
submodel.}

\item{Pi_bounds}{A numeric vector of lower and upper bounds for the
trial enrollment probabilities. The first element is the lower bound,
and the second element is the upper bound.}

\item{g}{A vector of estimated treatment probabilities,
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.}

\item{tau}{Returns from the \code{\link{learn_tau}} function.}
}
\value{
A \code{list} containing the following elements:
\item{pred}{Targeted estimates of trial enrollment probabilities;}
\item{A1}{Targeted estimates of trial enrollment probabilities under treatment;}
\item{A0}{Targeted estimates of trial enrollment probabilities under control.}
}
\description{
This function updates the trial enrollment probabilities
using the targeted maximum likelihood estimation (TMLE).
}
\keyword{TMLE}
