% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_R6.R
\name{tmle_R6}
\alias{tmle_R6}
\title{TMLE R6 Class}
\description{
TMLE R6 Class

TMLE R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TMLE-new}{\code{tmle_R6$new()}}
\item \href{#method-TMLE-create_folds}{\code{tmle_R6$create_folds()}}
\item \href{#method-TMLE-fit_regression}{\code{tmle_R6$fit_regression()}}
\item \href{#method-TMLE-target}{\code{tmle_R6$target()}}
\item \href{#method-TMLE-run}{\code{tmle_R6$run()}}
\item \href{#method-TMLE-inference}{\code{tmle_R6$inference()}}
\item \href{#method-TMLE-clone}{\code{tmle_R6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-new"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle_R6$new(data, W_nodes, A_node, Y_node, family, n_folds = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-create_folds"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-create_folds}{}}}
\subsection{Method \code{create_folds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle_R6$create_folds(n_eff, n_folds, strata_ids, seed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-fit_regression"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-fit_regression}{}}}
\subsection{Method \code{fit_regression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle_R6$fit_regression(
  method,
  folds,
  covariate_nodes,
  outcome_node,
  weight_node = NULL,
  subset = seq(nrow(self$data)),
  bound = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-target"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-target}{}}}
\subsection{Method \code{target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle_R6$target(g1W_bounds, g0W_bounds, browse = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-run"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle_R6$run(
  Q_method,
  g_method,
  family,
  g1W_bounds = NULL,
  g0W_bounds = NULL,
  seed = 123,
  browse = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-inference"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-inference}{}}}
\subsection{Method \code{inference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle_R6$inference(alpha = 0.05, browse = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TMLE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TMLE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle_R6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
