% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmle.R
\name{atmle}
\alias{atmle}
\title{Adaptive-TMLE for RCT-ATE}
\usage{
atmle(
  data,
  S_node,
  W_node,
  A_node,
  Y_node,
  controls_only,
  family,
  atmle_pooled = TRUE,
  r_loss = TRUE,
  theta_method = "glm",
  Pi_method = "glm",
  g_method = "glm",
  theta_tilde_method = "glm",
  Q_method = "glm",
  bias_working_model = "glmnet",
  pooled_working_model = "glmnet",
  g_rct,
  var_method = "ic",
  max_iter = 1,
  v_folds = 5,
  g_bound = c(0.025, 0.975),
  Pi_bound = c(0.025, 0.975),
  theta_bound = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} containing baseline covariates \eqn{W}, binary
treatment indicator \eqn{A} (\eqn{A=1} for active treatment),
outcome \eqn{Y}, and binary indicator of whether the observation is from the
randomized controlled trial \eqn{S=1} or from the external data \eqn{S=0}.}

\item{S_node}{The column indices of the \eqn{S} node in \code{data}.}

\item{W_node}{The column indices of the \eqn{W} node in \code{data}.}

\item{A_node}{The column indices of the \eqn{A} node in \code{data}.}

\item{Y_node}{The column indices of the \eqn{Y} node in \code{data}.}

\item{controls_only}{A logical indicating whether the external data has
only control-arm or both control-arm and treatment-arm.}

\item{family}{A character string specifying the family of the outcome
\eqn{Y}.
Currently only \code{"gaussian"} and \code{"binomial"} is supported.}

\item{atmle_pooled}{A logical indicating whether to use A-TMLE for the
pooled-ATE parameter also. If set to \code{FALSE}, use a regular TMLE.
Default is \code{TRUE}.}

\item{theta_method}{The method to estimate the nuisance function
\eqn{\theta(W,A)=\mathbb{E}(Y\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
or a \code{list} of \code{sl3} learners for super learner-based estimation.
Default is \code{"glmnet"}.}

\item{Pi_method}{The method to estimate the nuisance function
\eqn{\Pi(S\mid W,A)=\mathbb{P}(S=1\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
or a \code{list} of \code{sl3} learners for super learner-based estimation.}

\item{g_method}{The method to estimate the nuisance function
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
or a \code{list} of \code{sl3} learners for super learner-based estimation.}

\item{theta_tilde_method}{The method to estimate the nuisance function
\eqn{\tilde{\theta}(W,A)=\mathbb{E}(Y\mid W,A,S=1)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
or a \code{list} of \code{sl3} learners for super learner-based estimation.}

\item{Q_method}{The method to estimate the nuisance function
\eqn{Q(A,W)=\mathbb{E}(Y\mid W,A)}.
\code{"glm"} for main-term linear model, \code{"glmnet"} for lasso,
or a list of \code{sl3} learners for super learner-based estimation.}

\item{bias_working_model}{The working model for the bias estimand.
Either \code{"glmnet"} for lasso-based working model or \code{"HAL"} for highly adaptive
lasso-based working model. Default is \code{"glmnet"}.}

\item{pooled_working_model}{The working model for the pooled-ATE estimand.
Either \code{"glmnet"} for lasso-based working model or \code{"HAL"} for highly adaptive
lasso-based working model. Default is \code{"glmnet"}.}

\item{g_rct}{The probability of receiving the active treatment in the
randomized controlled trial.}

\item{var_method}{The method to estimate the variance of the A-TMLE
estimator. Either \code{"ic"} for influence curve-based variance estimator or
\code{"bootstrap"} for bootstrap-based variance estimator. Default is \code{"ic"}.}

\item{max_iter}{The maximum number of iterations for TMLE targeting of
\eqn{\Pi(S\mid W,A)=\mathbb{P}(S=1\mid W,A)}. Default is \code{1}.}

\item{v_folds}{The number of folds for cross-validation (whenever necessary).
Default is \code{5}.}

\item{verbose}{A logical indicating whether to print out the progress.
Default is \code{TRUE}.}
}
\value{
A \code{list} containing the following elements:
\item{ate}{The estimated average treatment effect;}
\item{ate_se}{The estimated standard error of the average treatment effect;}
\item{ate_ci}{The estimated \eqn{95\%} confidence interval for the average treatment
effect.}
}
\description{
Adaptive-TMLE (A-TMLE) for estimating the average treatment
effect based on combined randomized controlled trial data and real-world
data.
}
