% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmle_ate_fusion.R
\name{atmle_ate_fusion}
\alias{atmle_ate_fusion}
\title{Adaptive-TMLE for Data Integration}
\description{
Adaptive-TMLE for Data Integration

Adaptive-TMLE for Data Integration
}
\section{Super class}{
\code{atmle::TMLE} -> \code{A-TMLE for RCT + RWD}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-A-TMLE for RCT + RWD-new}{\code{atmle_ate_fusion$new()}}
\item \href{#method-A-TMLE for RCT + RWD-eval_Pi}{\code{atmle_ate_fusion$eval_Pi()}}
\item \href{#method-A-TMLE for RCT + RWD-eval_g_bar}{\code{atmle_ate_fusion$eval_g_bar()}}
\item \href{#method-A-TMLE for RCT + RWD-eval_Q_bar}{\code{atmle_ate_fusion$eval_Q_bar()}}
\item \href{#method-A-TMLE for RCT + RWD-eval_theta}{\code{atmle_ate_fusion$eval_theta()}}
\item \href{#method-A-TMLE for RCT + RWD-run_init_est}{\code{atmle_ate_fusion$run_init_est()}}
\item \href{#method-A-TMLE for RCT + RWD-fit_cate}{\code{atmle_ate_fusion$fit_cate()}}
\item \href{#method-A-TMLE for RCT + RWD-target_Pi}{\code{atmle_ate_fusion$target_Pi()}}
\item \href{#method-A-TMLE for RCT + RWD-target_tau}{\code{atmle_ate_fusion$target_tau()}}
\item \href{#method-A-TMLE for RCT + RWD-target_tau_A_tmle}{\code{atmle_ate_fusion$target_tau_A_tmle()}}
\item \href{#method-A-TMLE for RCT + RWD-target_tau_S_tmle}{\code{atmle_ate_fusion$target_tau_S_tmle()}}
\item \href{#method-A-TMLE for RCT + RWD-target_tau_relaxed}{\code{atmle_ate_fusion$target_tau_relaxed()}}
\item \href{#method-A-TMLE for RCT + RWD-target_beta_A_seq}{\code{atmle_ate_fusion$target_beta_A_seq()}}
\item \href{#method-A-TMLE for RCT + RWD-target_Pi_beta_S}{\code{atmle_ate_fusion$target_Pi_beta_S()}}
\item \href{#method-A-TMLE for RCT + RWD-inference}{\code{atmle_ate_fusion$inference()}}
\item \href{#method-A-TMLE for RCT + RWD-run}{\code{atmle_ate_fusion$run()}}
\item \href{#method-A-TMLE for RCT + RWD-clone}{\code{atmle_ate_fusion$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="atmle" data-topic="TMLE" data-id="create_folds"><a href='../../atmle/html/TMLE.html#method-TMLE-create_folds'><code>atmle::TMLE$create_folds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="atmle" data-topic="TMLE" data-id="fit_regression"><a href='../../atmle/html/TMLE.html#method-TMLE-fit_regression'><code>atmle::TMLE$fit_regression()</code></a></span></li>
<li><span class="pkg-link" data-pkg="atmle" data-topic="TMLE" data-id="target"><a href='../../atmle/html/TMLE.html#method-TMLE-target'><code>atmle::TMLE$target()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-new"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$new(
  data,
  S_node,
  W_nodes,
  A_node,
  Y_node,
  family,
  n_folds = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-eval_Pi"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-eval_Pi}{}}}
\subsection{Method \code{eval_Pi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$eval_Pi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-eval_g_bar"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-eval_g_bar}{}}}
\subsection{Method \code{eval_g_bar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$eval_g_bar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-eval_Q_bar"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-eval_Q_bar}{}}}
\subsection{Method \code{eval_Q_bar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$eval_Q_bar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-eval_theta"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-eval_theta}{}}}
\subsection{Method \code{eval_theta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$eval_theta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-run_init_est"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-run_init_est}{}}}
\subsection{Method \code{run_init_est()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$run_init_est(Q_method, Pi_bar_method, g_method, Q_bound)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-fit_cate"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-fit_cate}{}}}
\subsection{Method \code{fit_cate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$fit_cate(W, A, Y, g1W, theta, hal_args, parallel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-target_Pi"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-target_Pi}{}}}
\subsection{Method \code{target_Pi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$target_Pi(cate_fit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-target_tau"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-target_tau}{}}}
\subsection{Method \code{target_tau()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$target_tau(cate_fit, tau_A)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-target_tau_A_tmle"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-target_tau_A_tmle}{}}}
\subsection{Method \code{target_tau_A_tmle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$target_tau_A_tmle(phi_W, beta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-target_tau_S_tmle"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-target_tau_S_tmle}{}}}
\subsection{Method \code{target_tau_S_tmle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$target_tau_S_tmle(phi_WA, phi_W1, phi_W0, beta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-target_tau_relaxed"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-target_tau_relaxed}{}}}
\subsection{Method \code{target_tau_relaxed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$target_tau_relaxed(pseudo_outcome, pseudo_weights, phi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-target_beta_A_seq"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-target_beta_A_seq}{}}}
\subsection{Method \code{target_beta_A_seq()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$target_beta_A_seq(n_lambda)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-target_Pi_beta_S"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-target_Pi_beta_S}{}}}
\subsection{Method \code{target_Pi_beta_S()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$target_Pi_beta_S(n_lambda, max_iter, verbose)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-inference"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-inference}{}}}
\subsection{Method \code{inference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$inference(alpha = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-run"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$run(
  Q_method,
  Pi_bar_method,
  g_method,
  family = c("gaussian", "binomial"),
  A_cate_hal_args = list(max_degree = 3L, smoothness_orders = 1L, num_knots = 20L),
  S_cate_hal_args = list(max_degree = 3L, smoothness_orders = 1L, num_knots = 20L),
  target_method = "tmle",
  target_gwt = TRUE,
  Q_bound = NULL,
  max_iter = 50,
  n_lambda = 10,
  parallel = FALSE,
  verbose = TRUE,
  browse = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-A-TMLE for RCT + RWD-clone"></a>}}
\if{latex}{\out{\hypertarget{method-A-TMLE for RCT + RWD-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{atmle_ate_fusion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
