% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_data_and_nodes}
\alias{check_data_and_nodes}
\title{Sanity checks for input data and nodes}
\usage{
check_data_and_nodes(data, S_node, W_node, A_node, Y_node)
}
\description{
Sanity checks for input data and nodes
}
