% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_relaxed_hal.R
\name{fit_relaxed_hal}
\alias{fit_relaxed_hal}
\title{Fit relaxed HAL}
\usage{
fit_relaxed_hal(
  X,
  Y,
  family,
  weights = NULL,
  X_no_basis = NULL,
  X_unpenalized = NULL
)
}
\arguments{
\item{X}{A numeric matrix of covariates.}

\item{Y}{A numeric vector of outcomes.}

\item{family}{A character string specifying the family of the outcome.}

\item{weights}{A numeric vector of weights.}

\item{X_no_basis}{A numeric matrix of covariates that should not be used
to construct HAL basis.}

\item{X_unpenalized}{A numeric matrix of covariates that should not be
penalized by HAL.}
}
\value{
A list with the following components:
\item{beta}{A numeric vector of coefficients.}
\item{pred}{A numeric vector of predictions.}
\item{hal_basis_list}{A list of HAL basis.}
\item{x_basis}{A numeric matrix of basis.}
}
\description{
Fit relaxed HAL
}
