% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_relaxed_hal.R
\name{fit_relaxed_hal}
\alias{fit_relaxed_hal}
\title{Relaxed highly adaptive lasso}
\usage{
fit_relaxed_hal(
  X,
  Y,
  X_unpenalized,
  X_weak_penalized,
  X_weak_penalized_level,
  family,
  weights,
  relaxed,
  enumerate_basis_args = list(),
  fit_hal_args = list()
)
}
\arguments{
\item{X}{A numeric matrix of covariates.}

\item{Y}{A numeric vector of outcomes.}

\item{X_unpenalized}{A numeric matrix of covariates that should not be
penalized by HAL. Used for enforcing minimal working models.}

\item{family}{A character string specifying the family of the outcome.
Either "gaussian" or "binomial".}

\item{weights}{A numeric vector of weights.}

\item{enumerate_basis_args}{A list of arguments to pass to `enumerate_basis`
function in `hal9001`.}

\item{fit_hal_args}{A list of arguments to pass to `fit_hal` function in
`hal9001`.}
}
\value{
A list with the following components:
\item{beta}{A numeric vector of coefficients.}
\item{hal_basis_list}{A list of selected HAL basis.}
\item{x_basis}{The design matrix of selected basis.}
\item{pred}{A numeric vector of predictions.}
\item{hal_fit}{The output of `fit_hal`.}
}
\description{
First fit highly adaptive lasso, then fit a `glm` only on the
design matrix of the selected basis functions (with non-zero coefficients).
}
