% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_sl3_learners.R
\name{get_default_sl3_learners}
\alias{get_default_sl3_learners}
\title{Get default \code{sl3} learners}
\usage{
get_default_sl3_learners(family)
}
\arguments{
\item{family}{A character string specifying the family of the outcome.}
}
\description{
Get default \code{sl3} learners
}
