% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{get_eic_psi_tilde}
\alias{get_eic_psi_tilde}
\title{Efficient influence function for the pooled-ATE projection parameter}
\usage{
get_eic_psi_tilde(psi_tilde, g, theta, Y, A, n, weights)
}
\arguments{
\item{psi_tilde}{Results from \code{\link{learn_psi_tilde}}.}

\item{theta}{Results from \code{\link{learn_theta}}.}

\item{Y}{A vector of outcomes.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{n}{The number of observations.}

\item{weights}{A vector of (e.g. inverse-censoring) weights.}

\item{g_pred}{Results from \code{\link{learn_g}}.}
}
\value{
A vector of efficient influence function values.
}
\description{
Computes the efficient influence function for the pooled-ATE
projection parameter.
}
\keyword{internal}
