% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_g.R
\name{learn_g}
\alias{learn_g}
\title{Learn nuisance function: treatment mechanism}
\usage{
learn_g(S, W, A, g_rct, controls_only, method, v_folds, g_bounds)
}
\arguments{
\item{S}{A vector of study indicators, \eqn{S=1} for RCT, \eqn{S=0} for RWD.}

\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{g_rct}{A numeric of treatment probability in RCT.}

\item{controls_only}{A logical indicating whether the external data has only
control-arm observations.}

\item{method}{Learning method. \code{"glm"} for main-term linear model,
\code{"glmnet"} for lasso, or a \code{list} of \code{sl3} learners for
super learner-based estimation.}

\item{v_folds}{A numeric of number of folds for cross-validation
(when necessary).}

\item{g_bounds}{A numeric vector of lower and upper bounds for the
treatment mechanism. The first element is the lower bound, and the second
element is the upper bound.}
}
\value{
A numeric vector of estimated treatment probabilities.
}
\description{
Function to learn the treatment mechanism
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.
}
\details{
We provide some details on estimating the treatment mechanism
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}. Note that, we can rewrite the
treatment mechanism into the following form:
\deqn{g(A\mid W)=\mathbb{P}(A=1\mid S=1,W)\mathbb{P}(S=1\mid W)+
\mathbb{P}(A=1\mid S=0,W)\mathbb{P}(S=0\mid W).}
This form allows us to use the RCT randomization probability
(typically known) for the term \eqn{\mathbb{P}(A=1\mid S=1,W)}.
The rest of the terms are estimated using methods specified in \code{method}.
}
\examples{
# simulate data
set.seed(123)
n <- 500
S <- rbinom(n, 1, 0.5)
W1 <- rnorm(n); W2 <- rnorm(n); W <- cbind(W1, W2)
A <- numeric(n)
A[S == 1] <- rbinom(sum(S), 1, 0.67)
A[S == 0] <- rbinom(n-sum(S), 1, plogis(1.2*W1-0.9*W2))

# learn treatment mechanism
g <- learn_g(S, W, A, 0.67, FALSE, "glm", 5, c(0.01, 0.99))
}
\keyword{nuisance}
