% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_Pi.R
\name{learn_Pi}
\alias{learn_Pi}
\title{Learn nuisance function: trial enrollment mechanism}
\usage{
learn_Pi(g, A, Pi_bounds)
}
\arguments{
\item{g}{Returned list from `learn_g()` function.}
}
\value{
A \code{list} containing the following elements:
\item{pred}{The estimated trial enrollment probabilities;}
\item{A1}{The estimated trial enrollment probabilities under treatment;}
\item{A0}{The estimated trial enrollment probabilities under control.}
}
\description{
Function to learn the trial enrollment mechanism
\eqn{\Pi(S\mid W,A)=\mathbb{P}(S=1\mid W,A)}.
}
\keyword{nuisance}
