% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi_tmle.R
\name{Pi_tmle}
\alias{Pi_tmle}
\title{TMLE update of the trial enrollment probabilities}
\usage{
Pi_tmle(S, W, A, g, tau, Pi, controls_only, target_gwt = FALSE, weights = 1)
}
\arguments{
\item{S}{A vector of study indicators, \eqn{S=1} for RCT, \eqn{S=0} for RWD.}

\item{W}{A matrix of baseline covariates.}

\item{A}{A vector of treatment indicators, \eqn{A=1} for treatment-arm,
\eqn{A=0} for control-arm.}

\item{g}{A vector of estimated treatment probabilities,
\eqn{g(A\mid W)=\mathbb{P}(A=1\mid W)}.}

\item{tau}{Returns from the \code{\link{learn_tau}} function.}

\item{Pi}{Returns from the \code{\link{learn_Pi}} function.}

\item{controls_only}{A logical indicating whether the external data has only
control-arm observations.}
}
\value{
A \code{list} containing the following elements:
\item{pred}{Targeted estimates of trial enrollment probabilities;}
\item{A1}{Targeted estimates of trial enrollment probabilities under treatment;}
\item{A0}{Targeted estimates of trial enrollment probabilities under control.}
}
\description{
This function updates the trial enrollment probabilities
using the targeted maximum likelihood estimation (TMLE).
}
\keyword{TMLE}
